+++ tibiapy/client.py
@@ -13,7 +13,7 @@ from tibiapy import abc, AuctionDetails, AuctionFilters, CharacterBazaar
 from tibiapy.character import Character
 from tibiapy.creature import BoostedCreature
 from tibiapy.enums import Category, HouseOrder, HouseStatus, HouseType, NewsCategory, NewsType, VocationFilter
-from tibiapy.errors import Forbidden, NetworkError, SiteMaintenance
+from tibiapy.errors import Forbidden, NetworkError, SiteMaintenanceError
 from tibiapy.event import EventSchedule
 from tibiapy.forum import CMPostArchive, ForumAnnouncement, ForumBoard, ForumPost, ForumThread, ListedBoard
 from tibiapy.guild import Guild, GuildWars, ListedGuild
@@ -185,7 +185,7 @@ class Client:
             async with self.session.request(method, url, data=data, headers=headers) as resp:
                 log.info(f"{url} | {method} | {resp.status} {resp.reason}")
                 if "maintenance.tibia.com" in str(resp.url):
-                    raise SiteMaintenance("Tibia.com is down for maintenance.")
+                    raise SiteMaintenanceError("Tibia.com is down for maintenance.")
                 self._handle_status(resp.status)
                 response = RawResponse(resp, time.perf_counter()-init_time)
                 response.content = await resp.text()